Move field roots to state pkg: first attempt
