+++ beacon-chain/state/BUILD.bazel
@@ -6,6 +6,7 @@ go_library(
     srcs = [
         "cloners.go",
         "doc.go",
+        "field_roots.go",
         "field_trie.go",
         "getters.go",
         "setters.go",
@@ -31,10 +32,12 @@ go_library(
         "//beacon-chain/state/stateutil:go_default_library",
         "//proto/beacon/p2p/v1:go_default_library",
         "//shared/bytesutil:go_default_library",
+        "//shared/featureconfig:go_default_library",
         "//shared/hashutil:go_default_library",
         "//shared/htrutils:go_default_library",
         "//shared/params:go_default_library",
         "//shared/sliceutil:go_default_library",
+        "@com_github_dgraph_io_ristretto//:go_default_library",
         "@com_github_gogo_protobuf//proto:go_default_library",
         "@com_github_pkg_errors//:go_default_library",
         "@com_github_prysmaticlabs_eth2_types//:go_default_library",
