+++ beacon-chain/state/stateutil/BUILD.bazel
@@ -8,10 +8,8 @@ go_library(
         "attestations.go",
         "blocks.go",
         "reference.go",
-        "state_root.go",
         "trie_helpers.go",
         "validator_map_handler.go",
-        "validators.go",
     ],
     importpath = "github.com/prysmaticlabs/prysm/beacon-chain/state/stateutil",
     visibility = [
@@ -30,12 +28,10 @@ go_library(
         "//beacon-chain/core/state/stateutils:go_default_library",
         "//proto/beacon/p2p/v1:go_default_library",
         "//shared/bytesutil:go_default_library",
-        "//shared/featureconfig:go_default_library",
         "//shared/hashutil:go_default_library",
         "//shared/htrutils:go_default_library",
         "//shared/params:go_default_library",
         "//shared/trieutil:go_default_library",
-        "@com_github_dgraph_io_ristretto//:go_default_library",
         "@com_github_pkg_errors//:go_default_library",
         "@com_github_prysmaticlabs_eth2_types//:go_default_library",
         "@com_github_prysmaticlabs_ethereumapis//eth/v1alpha1:go_default_library",
