+++ beacon-chain/state/stateutil/trie_helpers_test.go
@@ -5,6 +5,7 @@ import (
 
 	types "github.com/prysmaticlabs/eth2-types"
 	ethpb "github.com/prysmaticlabs/ethereumapis/eth/v1alpha1"
+	"github.com/prysmaticlabs/prysm/beacon-chain/state"
 	"github.com/prysmaticlabs/prysm/beacon-chain/state/stateutil"
 	"github.com/prysmaticlabs/prysm/shared/bytesutil"
 	"github.com/prysmaticlabs/prysm/shared/hashutil"
@@ -16,7 +17,7 @@ import (
 
 func TestReturnTrieLayer_OK(t *testing.T) {
 	newState, _ := testutil.DeterministicGenesisState(t, 32)
-	root, err := stateutil.RootsArrayHashTreeRoot(newState.BlockRoots(), uint64(params.BeaconConfig().SlotsPerHistoricalRoot), "BlockRoots")
+	root, err := state.RootsArrayHashTreeRoot(newState.BlockRoots(), uint64(params.BeaconConfig().SlotsPerHistoricalRoot), "BlockRoots")
 	require.NoError(t, err)
 	blockRts := newState.BlockRoots()
 	roots := make([][32]byte, 0, len(blockRts))
@@ -30,13 +31,13 @@ func TestReturnTrieLayer_OK(t *testing.T) {
 
 func TestReturnTrieLayerVariable_OK(t *testing.T) {
 	newState, _ := testutil.DeterministicGenesisState(t, 32)
-	root, err := stateutil.ValidatorRegistryRoot(newState.Validators())
+	root, err := state.ValidatorRegistryRoot(newState.Validators())
 	require.NoError(t, err)
 	hasher := hashutil.CustomSHA256Hasher()
 	validators := newState.Validators()
 	roots := make([][32]byte, 0, len(validators))
 	for _, val := range validators {
-		rt, err := stateutil.ValidatorRoot(hasher, val)
+		rt, err := state.ValidatorRoot(hasher, val)
 		require.NoError(t, err)
 		roots = append(roots, rt)
 	}
@@ -74,7 +75,7 @@ func TestRecomputeFromLayer_VariableSizedArray(t *testing.T) {
 	hasher := hashutil.CustomSHA256Hasher()
 	roots := make([][32]byte, 0, len(validators))
 	for _, val := range validators {
-		rt, err := stateutil.ValidatorRoot(hasher, val)
+		rt, err := state.ValidatorRoot(hasher, val)
 		require.NoError(t, err)
 		roots = append(roots, rt)
 	}
@@ -95,11 +96,11 @@ func TestRecomputeFromLayer_VariableSizedArray(t *testing.T) {
 	require.NoError(t, newState.UpdateValidatorAtIndex(types.ValidatorIndex(changedIdx[0]), changedVals[0]))
 	require.NoError(t, newState.UpdateValidatorAtIndex(types.ValidatorIndex(changedIdx[1]), changedVals[1]))
 
-	expectedRoot, err := stateutil.ValidatorRegistryRoot(newState.Validators())
+	expectedRoot, err := state.ValidatorRegistryRoot(newState.Validators())
 	require.NoError(t, err)
 	roots = make([][32]byte, 0, len(changedVals))
 	for _, val := range changedVals {
-		rt, err := stateutil.ValidatorRoot(hasher, val)
+		rt, err := state.ValidatorRoot(hasher, val)
 		require.NoError(t, err)
 		roots = append(roots, rt)
 	}
